requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
Werkzeug==2.2.2
zipp==3.11.0
redis~=4.4.2
uwsgi
PyYAML==6.0